test(pairing): split command unit coverage
